+++ source/isaaclab/docs/CHANGELOG.rst
@@ -35,6 +35,18 @@ Changed
   * ``write_joint_friction_to_sim`` → ``write_joint_friction_coefficient_to_sim``
 
 
+0.34.7 (2025-03-04)
+~~~~~~~~~~~~~~~~~~~
+
+Added
+^^^^^
+
+* Added methods inside the :class:`omni.isaac.lab.assets.Articulation` class to set the joint
+  position and velocity for the articulation. Previously, the joint position and velocity could
+  only be set using the :meth:`omni.isaac.lab.assets.Articulation.write_joint_state_to_sim` method,
+  which didn't allow setting the joint position and velocity separately.
+
+
 0.34.6 (2025-03-02)
 ~~~~~~~~~~~~~~~~~~~